git-svn-id: https://svn.r-project.org/R/branches/R-BIGSEXP@68681 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/tre/tre-mem.c
@@ -111,8 +111,8 @@ tre_mem_alloc_impl(tre_mem_t mem, int provided, void *provided_block,
 	    block_size = size * 8;
 	  else
 	    block_size = TRE_MEM_BLOCK_SIZE;
-	  DPRINT(("tre_mem_alloc: allocating new %u byte block\n",
-		  (unsigned int)block_size));
+	  DPRINT(("tre_mem_alloc: allocating new %lu byte block\n",
+		  (unsigned long)block_size));
 	  l = xmalloc(sizeof(*l));
 	  if (l == NULL)
 	    {
